#162 Change: コメントに関する処理をtodos.jsからcomment.jsに移行
